+++ dbentry/base/views.py
@@ -5,7 +5,6 @@ from django.contrib.admin import AdminSite
 from django.contrib.auth.mixins import UserPassesTestMixin
 from django.forms import Form
 from django.http import HttpRequest, HttpResponse
-from formtools.wizard.views import SessionWizardView
 
 from dbentry.sites import miz_site
 
@@ -62,49 +61,6 @@ class OptionalFormView(views.generic.FormView):
         else:
             # This view has a form_class set, but the form is invalid.
             return self.form_invalid(form)
-
-
-class FixedSessionWizardView(SessionWizardView):
-    """Subclass of SessionWizardView that fixes some quirks."""
-
-    def get_context_data(self, **kwargs: Any) -> dict:
-        # SessionWizardView expects 'form' as required positional argument.
-        # TODO: this isn't the case anymore, SessionWizardView plays nice with
-        #  form as a kwarg.
-        return super().get_context_data(kwargs.pop('form', None), **kwargs)
-
-    def get_form(
-            self,
-            step: Optional[int] = None,
-            data: Optional[dict] = None,
-            files: Optional[dict] = None
-    ) -> Form:
-        """
-        WizardView.get_form overrides any alterations to the form kwargs
-        made in self.get_form_kwargs:
-
-            kwargs = self.get_form_kwargs(step)
-            kwargs.update({
-                'data': data,
-                'files': files,
-                'prefix': self.get_form_prefix(step, form_class),
-                'initial': self.get_form_initial(step),
-            })
-            Where 'step', 'data' and 'files' are get_form's arguments.
-
-        If we wanted to keep changes to the request payload made in
-        get_form_kwargs, we need to call super with explicit data (and files)
-        kwargs derived from get_form_kwargs.
-        """
-        # TODO: it doesn't look like any views using this view class use
-        #  get_form_kwargs to override data/files - so this 'fix' here might
-        #  not be needed.
-        if step is None:
-            step = self.steps.current
-        kwargs = self.get_form_kwargs(step)
-        data = data or kwargs.get('data', None)
-        files = files or kwargs.get('files', None)
-        return super().get_form(step, data, files)
 
 
 class SuperUserOnlyMixin(UserPassesTestMixin):